+++ ios/Classes/CleverTapPlugin.m
@@ -245,9 +245,7 @@ static NSDateFormatter *dateFormatter;
         result(nil);
     else if ([@"deleteNotificationChannel" isEqualToString:call.method])
         result(nil);
-    else if ([@"setBaiduPushToken" isEqualToString:call.method])
-        result(nil);
-    else if ([@"setHuaweiPushToken" isEqualToString:call.method])
+    else if ([@"pushRegistrationToken" isEqualToString:call.method])
         result(nil);
     else if ([@"setLibrary" isEqualToString:call.method])
         [self setLibrary:call withResult:result];