+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 ## CHANGE LOG
 
+Version 3.3.0 *(27 March 2025)*
+-------------------------------------------
+**What's new**
+* **[Android Platform]**
+  * Supports [CleverTap Android SDK v7.3.1](https://github.com/CleverTap/clevertap-android-sdk/blob/master/docs/CTCORECHANGELOG.md#version-731-march-27-2025).
+  * Adds new API to register push tokens for providers other than FCM
+    * `pushRegistrationToken(String value, Map<String, String> pushType)`: Register the token for the specified PushType. Refer [here](doc/Usage.md#registering-fcm-baidu-huawei-token) for example usage
+
+**Breaking API Changes**
+  - **Removes:** The older token registration APIs have been removed to make push providers injectable. The removed APIs include:
+    - `setBaiduPushToken(String value)`
+    - `setHuaweiPushToken(String value)`
+  
+
 Version 3.2.0 *(3 March 2025)*
 -------------------------------------------
 > ⚠️ **NOTE**